Fix small domain bug
git-svn-id: svn+ssh://keith.cs.st-and.ac.uk/mnt/raid/svn/repositories/minion-private/public/trunk@629 477b24dd-e91e-0410-aa49-980f3df0729b

+++ minion/constraint_setup.h
@@ -37,7 +37,7 @@ inline bool special_limit_propogate_static_queue(const set<Constraint*>& constra
 	while(!propogate_trigger_list.empty())
 	{
 	  TriggerRange t = propogate_trigger_list.back();
-	  short data_val = t.data;
+	  int data_val = t.data;
 	  propogate_trigger_list.pop_back();
 	  
 	  for(Trigger* it = t.begin(); it != t.end(); it++)
